A sketch operation to form `[n][k]` from `[n]`.
There's some nice things about this function. The `[n][n]` Gram matrix
is never explicitly formed. There are things that could make it far
nice however: if we utilized a CBRNG, we wouldn't have to form the
Gaussian noise matrix beforehand -- we could compute it on the fly,
which would likely result in some nice performance gains.